Split up handlers.py into smaller files
